requirements.txt
google-generativeai
streamlit
langchain
langgraph
python-dotenv
google-cloud-storage
google-cloud-logging
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery

